Move remove participant from team to participants_functionality
